requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
firecrawl-py>=0.0.16